clean some code
git-svn-id: svn://anonsvn.kde.org/home/kde/trunk/kdemultimedia/kmidi@33492 283d02a7-25f6-0310-bc7c-ecb5cbfe19da

+++ linux_a.c
@@ -48,7 +48,7 @@ extern void b_out(int fd, int *buf, int ocount);
 
 static int open_output(void); /* 0=success, 1=warning, -1=fatal error */
 static void close_output(void);
-static void output_data(int32 *buf, int32 count);
+static void output_data(int32 *buf, uint32 count);
 static void flush_output(void);
 static void purge_output(void);
 
@@ -232,12 +232,12 @@ int current_sample_count()
   return samples;
 }
 
-static void output_data(int32 *buf, int32 count)
+static void output_data(int32 *buf, uint32 count)
 {
   int ocount;
 
   if (!(dpm.encoding & PE_MONO)) count*=2; /* Stereo samples */
-  ocount = count;
+  ocount = (int)count;
 
   if (ocount) {
     if (dpm.encoding & PE_16BIT)
